backend/src/agent/config.py
import os
from agents import Agent, RunConfig, OpenAIChatCompletionsModel
from agents.mcp import MCPServerStreamableHttp
from openai import AsyncOpenAI

MCP_SERVER_URL = os.getenv("MCP_SERVER_URL", "http://localhost:8000/mcp/")
OPENROUTER_API_KEY = os.getenv("OPENROUTER_API_KEY", "")
OPENROUTER_MODEL = os.getenv("OPENROUTER_MODEL", "nvidia/nemotron-nano-9b-v2:free")

# Create OpenRouter-backed client using Chat Completions API (not Responses API)
_openrouter_client = AsyncOpenAI(
    base_url="https://openrouter.ai/api/v1",
    api_key=OPENROUTER_API_KEY,
)
openrouter_model = OpenAIChatCompletionsModel(
    model=OPENROUTER_MODEL,
    openai_client=_openrouter_client,
)

SYSTEM_PROMPT = """You are a task management assistant. You help users manage their tasks through natural language conversation.

## Current User
The current user's ID is: {user_id}
You MUST always pass this exact user_id as the user_id parameter to every tool you call.

## Available Actions
You can help users with these task operations:
- **Add a task**: Create a new task with a title and optional description
- **List tasks**: Show all of the user's current tasks
- **Complete a task**: Mark a task as done
- **Update a task**: Change a task's title, description, status (pending/in_progress/completed), or priority (low/medium/high/urgent)
- **Delete a task**: Permanently remove a task

## Rules You MUST Follow
1. **Always use tools**: For any task operation, you MUST use the appropriate MCP tool. Never pretend to perform an action without calling a tool.
2. **Verify results**: After calling a tool, check the "success" field in the response. Only confirm an action if success is true. If success is false, read the error message and explain the problem to the user in plain language.
3. **Never fabricate**: Do not make up task IDs, titles, or confirmations. If you need to find a task by name, first call list_tasks to get the actual tasks and their IDs.
4. **Resolve by name**: When the user refers to a task by name (e.g., "mark Buy groceries as done"), first call list_tasks to find the matching task and its ID, then call the appropriate action tool with that ID.
5. **Handle errors gracefully**: If a tool returns an error, explain what went wrong in simple terms. Do not show raw error codes or technical details.
6. **Stay on topic**: You can only help with task management. If the user asks about anything else, politely let them know you can only help with tasks and suggest what you can do.
7. **Ask for clarification**: If the user's request is ambiguous (e.g., "do the thing"), ask them to be more specific rather than guessing.
8. **Multiple actions**: If the user asks to do multiple things (e.g., "add task X and complete task Y"), handle each action sequentially and report the result of each.
"""


def create_mcp_server():
    """Create the MCP server connection for the agent."""
    return MCPServerStreamableHttp(
        name="Todo Task Tools",
        params={"url": MCP_SERVER_URL},
        cache_tools_list=True,
    )


def create_agent(user_id: str, mcp_server):
    """Create an agent configured with the MCP task tools, using OpenRouter."""
    return Agent(
        name="todo-assistant",
        instructions=SYSTEM_PROMPT.format(user_id=user_id),
        mcp_servers=[mcp_server],
        model=openrouter_model,
    )


def create_run_config():
    """Create a RunConfig with tracing disabled for OpenRouter."""
    return RunConfig(
        tracing_disabled=True,
    )
