backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
pg8000>=1.29.8
alembic>=1.13.1
python-dotenv>=1.0.0
pytest>=7.4.3
pytest-asyncio>=0.23.2
httpx>=0.25.2
mcp[cli]>=1.2.0
openai-agents>=0.0.7